refactor: Extract audio logic into AudioController
Separate audio management logic from UI code for better maintainability
and single responsibility principle.

Changes:
- Create AudioController.kt to encapsulate all AudioManager and
  LoudnessEnhancer logic
- Move audio methods from ActivityMain to AudioController:
  - setInitialVolume()
  - applyBoost(level: Int)
  - applyVolume(level: Int)
  - restartAudioPlayback()
  - getOutputDeviceInfo()
  - release()
- Update ActivityMain to delegate audio operations to AudioController
- Preserve all existing logic, formatting, and behavior
